Limiting the balance decimal

+++ backend/seeder.js
@@ -44,7 +44,7 @@ const importData = async () => {
       public_address uuid NOT NULL DEFAULT uuid_generate_v4(),
       user_id uuid DEFAULT uuid_generate_v4() NOT NULL REFERENCES users(id),
       coin_id INT NOT NULL REFERENCES coins(id),
-      balance FLOAT NOT NULL CHECK (balance >= 0),
+      balance NUMERIC(20, 5) NOT NULL CHECK (balance >= 0),
       created_at DATE NOT NULL DEFAULT CURRENT_DATE,
       used BOOLEAN NOT NULL DEFAULT true
     );
